Finished BMI now working on the User Avatar
