+++ src/main/java/org/launchcode/uTrain/controllers/UserController.java
@@ -117,15 +117,17 @@ public class UserController {
         return "user/profile";
     }
 
-//    @GetMapping("bmi")
-//    public String bmiCalc(HttpServletRequest request, Model model) {
-//
-//        User user = (User) getUserFromSession(request.getSession());
-//
-//        model.addAttribute("title", "BMI Calculator");
-//        model.addAttribute("user", user);
-//        return "bmi";
-//    }
+    @GetMapping("bmi")
+    public String bmiCalc(HttpServletRequest request, Model model) {
 
+        User user = (User) getUserFromSession(request.getSession());
+
+        model.addAttribute("title", "BMI Calculator");
+        model.addAttribute("bmi1", user.getUserDetail().getBodyMassIndex());
+        model.addAttribute("user", user);
+        model.addAttribute("loggedIn", true);
+
+        return "user/bmi";
+    }
 
 }
