--- src/main/java/org/launchcode/uTrain/data/BmiRepository.java
@@ -1,4 +1,0 @@
-package org.launchcode.uTrain.data;
-
-public interface BmiRepository {
-}
