--- src/main/resources/templates/bmi/bmi.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org/">
-    <head th:replace="fragments :: head">
-    </head>
-    <body>
-        <div th:replace="fragments :: page-header"></div>
-
-
-        <h1>Your height is: </h1>
-        <h1 th:text="${user.userDetail.height}"></h1>
-        <h2>Your weight is: </h2>
-        <h2 th:text="${user.userDetail.weight}"></h2>
-        <h3>Your BMI is: </h3>
-        <h3 th:text="${bmiReturn}"></h3>
-        <form name="calculator">
-            <table>
-                <tr>
-                    <td>
-                        <input type="text" name="display" id="display"
-                        disabled>
-                    </td>
-                </tr>
-                <tr>
-                    <td><input type="button" name="one" value="1" onclick="calculator.display.value += '1'"> </td>
-                    <td><input type="button" name="two" value="2" onclick="calculator.display.value += '2'"> </td>
-                    <td><input type="button" name="three" value="3" onclick="calculator.display.value += '3'"> </td>
-                    <td><input type="button" name="plus" value="+" onclick="calculator.display.value += '+'"> </td>
-                </tr>
-                <tr>
-                    <td><input type="button" name="four" value="4" onclick="calculator.display.value += '4'"> </td>
-                    <td><input type="button" name="five" value="5" onclick="calculator.display.value += '5'"> </td>
-                    <td><input type="button" name="six" value="6" onclick="calculator.display.value += '6'"> </td>
-                    <td><input type="button" name="minus" value="-" onclick="calculator.display.value += '-'"> </td>
-                </tr>
-                <tr>
-                    <td><input type="button" name="seven" value="7" onclick="calculator.display.value += '7'"> </td>
-                    <td><input type="button" name="eight" value="8" onclick="calculator.display.value += '8'"> </td>
-                    <td><input type="button" name="nine" value="9" onclick="calculator.display.value += '9'"> </td>
-                    <td><input type="button" name="times" value="*" onclick="calculator.display.value += '*'"> </td>
-                    <td><input type="button" name="period" value="." onclick="calculator.display.value += '.'"> </td>
-                </tr>
-                <tr>
-                    <td><input type="button" name="clear" value="c" onclick="calculator.display.value = ''"> </td>
-                    <td><input type="button" name="zero" value="0" onclick="calculator.display.value += '0'"> </td>
-                    <td><input type="button" name="doit" value="=" onclick="calculator.display.value = eval(calculator.display.value)"> </td>
-                    <td><input type="button" name="divide" value="/" onclick="calculator.display.value += '/'"> </td>
-                </tr>
-            </table>
-        </form>
-
-        <footer th:replace="fragments :: footer"></footer>
-
-    </body>
-</html>
