+++ src/main/resources/templates/fragments.html
@@ -19,7 +19,9 @@
             <img src="https://i.ibb.co/mFszRzy/IMG-1539.png" alt="logo" border="0" width="60" height="51">
         </a>
     </li>
-    <li><a href="/bmi/bmi">BMI Calculator</a> </li>
+    <li th:if="${loggedIn}"><a href="/user/bmi">BMI Calculator</a> </li>
+    <li th:unless="${loggedIn}"><a href="/bmiLoggedOut">BMI</a> </li>
+
     <li><a href="/company">About Us</a></li>
 
     <li th:if="${loggedIn}"><a  href="/user/index">Home</a></li>
