+++ src/main/resources/templates/user/bmi.html
@@ -1,0 +1,49 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org/">
+    <head th:replace="fragments :: head">
+    </head>
+    <body>
+        <div th:replace="fragments :: page-header"></div>
+
+
+        <h1>Your height is: </h1>
+        <h1 th:text="${user.userDetail.height}"></h1>
+        <h2>Your weight is: </h2>
+        <h2 th:text="${user.userDetail.weight}"></h2>
+        <h3>Your BMI is: </h3>
+        <h3 th:text="${bmi1}"></h3>
+        <div name="bmiTable" style="float: right">
+            <table>
+                <tr>
+                    <td class="titleTable">
+                        BMI
+                    </td>
+                </tr>
+                <tr>
+                    <td class="subTitleTable">Classification </td>
+                    <td class="subTitleTable">BMI Target</td>
+                </tr>
+                <tr>
+                    <td class="miniTitleTable">Underweight</td>
+                    <td class="numTable">Below 18.5</td>
+                </tr>
+                <tr>
+                    <td class="miniTitleTable">Healthy</td>
+                    <td>Between 18.5 -24.9</td>
+
+                </tr>
+                <tr>
+                    <td class="miniTitleTable">Overweight</td>
+                    <td>25 - 29.9</td>
+                </tr>
+                <tr>
+                    <td class="miniTitleTable">Obese</td>
+                    <td>Over 30</td>
+                </tr>
+            </table>
+        </div>
+
+        <footer th:replace="fragments :: footer"></footer>
+
+    </body>
+</html>
